create_postal2senkyoku_json.js　をクラスごとにファイル分割
